# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Guardian
  version: 0.1.0
  description: Guardian API docs
  contact:
    email: contact.tharudealwis@gmail.com
servers:
  - url: http://localhost:2699/api/v1
    description: Local development server
...
paths:
...
  /reports/witness/{reportId}/{witnessId}:
    delete:
      tags:
        - Reports
      summary: Allows officers to delete report witnesses
      operationId: deleteReportWitness
      description: >
        Delete a witness from a report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reportId
          schema:
            type: integer
          required: true
          description: Numeric ID of the report to modify
        - in: path
          name: witnessId
          schema:
            type: integer
          required: true
          description: Numeric ID of the witness to delete

      responses:
        "204":
          description: Deleted Confirmation
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalErrorResponse"
...
  /lost-articles/personal-details/{lostArticleId}/{personalDetailsId}:
    delete:
      tags:
        - Lost Articles
      summary: Delete personal details to a article by id
      operationId: deleteLostArticlePersonalDetails
      description: >
        Delete the personal details from a lost article report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: lostArticleId
          schema:
            type: integer
          required: true
          description: Numeric ID of the lost article to modify
        - in: path
          name: personalDetailsId
          schema:
            type: integer
          required: true
          description: Numeric ID of the personal details to delete
      responses:
        "204":
          description: Deleted Confirmation
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: apiKey
      in: header
      name: refresh-token
...
  schemas:
    Response:
      type: object
      required:
        - status
        - data
        - message
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          description: The response data
        message:
          type: string
          example: ""
...
    UnauthorizedErrorResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            status:
              type: string
              example: "error"
            data:
              type: object
              properties:
                code:
                  type: integer
                  example: 401
                message:
                  type: string
                  example: ""
                data:
                  type: object
                  example: {}
    InternalErrorResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            message:
              type: string
              example: Internal Server Error
            status:
              type: string
              example: "error"